Issues
'Bio' field in User model #7
'Job Title' field in User model #9
'Team' field in User model #10

Solution

Add bio, team and job title to the User model. Each is set as a Charfield. Bio would suit a TesxtArea but can't be used in a form due to validation (set required=False, as may not be necessary on all sites, but can be changed on a site by site basis. Also added this country as well)