Different solution to implement Sections as outlined in #1.
The types are separated into Pill/Tab/Accordion Blocks with separate templates.

There is an option for basic layout styling within each block (if applicable).

Each of these blocks belong to a 'group' called 'Section Blocks' so that in the CoreBlock they are grouped together. Making it easier to find the block you need.